abstraction v1 for scraping complete
