+++ bigbang/utils.py
@@ -1,12 +1,170 @@
 """
 Miscellaneous utility functions used in other modules.
 """
-
+import os
+import time
+import logging
 import glob
+import yaml
+from pathlib import Path
 from typing import Dict, List, Optional, Tuple, Union
+from bs4 import BeautifulSoup
+import requests
 
 import networkx as nx
 import pandas as pd
+
+from config.config import CONFIG
+
+filepath_auth = CONFIG.config_path + "authentication.yaml"
+directory_project = str(Path(os.path.abspath(__file__)).parent.parent)
+logging.basicConfig(
+    filename=directory_project + "/scraping.log",
+    filemode="w",
+    level=logging.INFO,
+    format="%(asctime)s %(message)s",
+)
+logger = logging.getLogger(__name__)
+
+
+def get_website_content(
+    url: str,
+    session: Optional[requests.Session] = None,
+) -> Union[str, BeautifulSoup]:
+    """
+    Get HTML code from website
+
+    Note
+    ----
+    Servers don't like it when one is sending too many requests from same
+    ip address in short period of time. Therefore we need to:
+        a) catch 'requests.exceptions.RequestException' errors
+            (includes all possible errors to be on the safe side),
+        b) safe intermediate results,
+        c) continue where we left off at a later stage.
+    """
+    # TODO: include option to change BeautifulSoup args
+    try:
+        if session is None:
+            sauce = requests.get(url)
+            assert sauce.status_code == 200
+            soup = BeautifulSoup(sauce.content, "html.parser")
+        else:
+            sauce = session.get(url)
+            soup = BeautifulSoup(sauce.text, "html.parser")
+        return soup
+    except requests.exceptions.RequestException as e:
+        if "A2=" in url:
+            # if URL of mboxMessage
+            logger.info(f"{e} for {url}.")
+            return "RequestException"
+        else:
+            SystemExit()
+
+
+def set_website_preference_for_header(
+    url_pref: str,
+    session: requests.Session,
+) -> requests.Session:
+    """
+    Set the 'Email Headers' of the 'Archive Preferences' for the auth session
+    to 'Show All Headers'. Otherwise only a restricted list of header fields is
+    shown.
+    """
+    url_archpref = url_pref + "&TAB=2"
+    payload = {
+        "Email Headers": "b",
+    }
+    session.post(url_archpref, data=payload)
+    return session
+
+
+def get_auth_session(
+    url_login: str, username: str, password: str
+) -> requests.Session:
+    """
+    Create AuthSession.
+
+    There are three ways to create an AuthSession:
+        - parse username & password directly into method
+        - create a /bigbang/config/authentication.yaml file that contains keys
+        - type then into terminal when the method 'get_login_from_terminal'
+            is raised
+    """
+    if os.path.isfile(filepath_auth):
+        # read from /config/authentication.yaml
+        with open(filepath_auth, "r") as stream:
+            auth_key = yaml.safe_load(stream)
+        username = auth_key["username"]
+        password = auth_key["password"]
+    else:
+        # ask user for login keys
+        username, password = get_login_from_terminal(username, password)
+
+    if username is None or password is None:
+        # continue without authentication
+        return None
+    else:
+        # Start the AuthSession
+        session = requests.Session()
+        # Create the payload
+        payload = {
+            "LOGIN1": "",
+            "Y": username,
+            "p": password,
+            "X": "",
+        }
+        # Post the payload to the site to log in
+        session.post(url_login, data=payload)
+        return session
+
+
+def get_login_from_terminal(
+    username: Union[str, None],
+    password: Union[str, None],
+    file_auth: str = directory_project + "/config/authentication.yaml",
+) -> Tuple[Union[str, None]]:
+    """
+    Get login key from user during run time if 'username' and/or 'password' is 'None'.
+    Return 'None' if no reply within 15 sec.
+    """
+    if username is None or password is None:
+        record = True
+    else:
+        record = False
+    if username is None:
+        username = ask_for_input("Enter your Email: ")
+    if password is None:
+        password = ask_for_input("Enter your Password: ")
+    if record and isinstance(username, str) and isinstance(password, str):
+        loginkey_to_file(username, password, file_auth)
+    return username, password
+
+
+def ask_for_input(request: str) -> Union[str, None]:
+    timeout = 15
+    end_time = time.time() + timeout
+    while time.time() < end_time:
+        reply = input(request)
+        try:
+            assert isinstance(reply, str)
+            break
+        except Exception:
+            reply = None
+            continue
+    return reply
+
+
+def loginkey_to_file(
+    username: str,
+    password: str,
+    file_auth: str,
+) -> None:
+    """Safe login key to yaml"""
+    file = open(file_auth, "w")
+    file.write(f"username: '{username}'\n")
+    file.write(f"password: '{password}'")
+    file.close()
 
 
 def get_paths_to_files_in_directory(
