+++ tests/webscraping/test_listserv.py
@@ -13,6 +13,7 @@ from bigbang.listserv import (
     ListservList,
     ListservMessageParser,
 )
+from bigbang.utils import get_login_from_terminal
 from config.config import CONFIG
 
 dir_temp = tempfile.gettempdir()
@@ -139,6 +140,8 @@ class TestListservList:
                 "weeks": 4,
                 "fields": "header",
             },
+            url_login="https://list.etsi.org/scripts/wa.exe?LOGON",
+            url_pref="https://list.etsi.org/scripts/wa.exe?PREF",
             login=auth_key,
         )
         assert mlist.messages[0]["from"] == "iceeng-10"
@@ -211,6 +214,8 @@ class TestListservArchive:
                 "weeks": 4,
                 "fields": "header",
             },
+            url_login="https://list.etsi.org/scripts/wa.exe?LOGON",
+            url_pref="https://list.etsi.org/scripts/wa.exe?PREF",
             login=auth_key_mock,
             instant_save=False,
             only_mlist_urls=False,
@@ -247,11 +252,11 @@ class TestListservArchive:
             Path(filepath).unlink()
 
 
-@mock.patch("bigbang.listserv.ask_for_input", return_value="check")
+@mock.patch("bigbang.utils.ask_for_input", return_value="check")
 def test__get_login_from_terminal(input):
     """test if login keys will be documented"""
     file_auth = dir_temp + "/authentication.yaml"
-    _, _ = listserv.get_login_from_terminal(
+    _, _ = get_login_from_terminal(
         username=None, password=None, file_auth=file_auth
     )
     f = open(file_auth, "r")
